Initial commit of file/JSON/movement test
The 10,000 foot goal of this exercise is to create named files within a new
directory, each with JSON data, and move them from one folder to the next
while versioning up one number, as well as changing the file names to GUIDs.

Currently, the program makes initial directory, and makes a new instance of
RandomJson class.

More to come!