Sript now updates JSON version key in each file
Though, it only appends to the end of each file in Modified. The next
fix will be to rewrite the contents instead of appending.

To help with this I, had to ask a senior developer friend of mine as I will admit,
I was hitting some roadblocks. He guided me and showed me that my to_s in RandomJSON
was creating incorrect JSON formatting. I changed this to be a hash, using generate, and
it wrote correct JSON output.

Now the hang up is actually in the writing of the file. More to come!

Also, I changed the number of output files in the script so that I'm not having to
sift through 20 files during testing. This will be changed back to 20